Added functionality to add extension pointers

+++ src/main/java/de/tobiasrick/pointerdiagram/pointer/ExtensionPointer.java
@@ -9,11 +9,11 @@ import javafx.scene.paint.Color;
  * @author Tobias Rick
  */
 public class ExtensionPointer extends Pointer{
-    public ExtensionPointer(GraphicsContext gc, int x1, int y1, int x2, int y2) {
+    public ExtensionPointer(GraphicsContext gc, double x1, double y1, double x2, double y2) {
         super(gc, x1, y1, x2, y2);
     }
 
-    public ExtensionPointer(GraphicsContext gc, int x1, int y1, int x2, int y2, Color strokeColor, Color fillColor) {
+    public ExtensionPointer(GraphicsContext gc, double x1, double y1, double x2, double y2, Color strokeColor, Color fillColor) {
         super(gc, x1, y1, x2, y2, strokeColor, fillColor);
     }
 }
